Don't commit to CVS when installing

+++ makesql.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 # guess version
-ver=`egrep "^castor" debian/changelog | head -1 | awk '{print $2}' | sed 's/)//' | sed 's/(//'`
+ver=`grep "^castor" debian/changelog | head -1 | awk '{print $2}' | sed 's/)//' | sed 's/(//'`
 tag=`echo $ver | sed 's/\./_/g' | sed 's/-/_/g'`
 
 # generate creation scripts for the db-based CASTOR components
@@ -14,17 +14,19 @@ cd dlf/scripts
 cd ../..
 
 # commit and tag in CVS
-sqlscripts='castor/db/castor_oracle_create.sql castor/db/castor_oracle_create.sqlplus castor/repack/repack_oracle_create.sql castor/repack/repack_oracle_create.sqlplus castor/vdqm/vdqm_oracle_create.sql castor/vdqm/vdqm_oracle_create.sqlplus dlf/scripts/oracle/dlf_oracle_create.sql dlf/scripts/oracle/dlf_oracle_create.sqlplus'
-echo Running CVS diff...
-cvs diff $sqlscripts > /tmp/cvsdiffs
-less /tmp/cvsdiffs
-echo 'Commit and tag in CVS? (y/n)'
-read ask
-if [ "$ask" == "y" ]; then
-  cvs commit -m "Regenerated creation scripts" $sqlscripts
-  cvs tag v$tag $sqlscripts
-  echo All SQL creation scripts for release $ver generated and tagged
+if [ $# == 0 ]; then
+  sqlscripts='castor/db/castor_oracle_create.sql castor/db/castor_oracle_create.sqlplus castor/repack/repack_oracle_create.sql castor/repack/repack_oracle_create.sqlplus castor/vdqm/vdqm_oracle_create.sql castor/vdqm/vdqm_oracle_create.sqlplus dlf/scripts/oracle/dlf_oracle_create.sql dlf/scripts/oracle/dlf_oracle_create.sqlplus'
+  echo Running CVS diff...
+  cvs diff $sqlscripts > /tmp/cvsdiffs
+  less /tmp/cvsdiffs
+  echo 'Commit and tag in CVS? (y/n)'
+  read ask
+  if [ "$ask" == "y" ]; then
+    cvs commit -m "Regenerated creation scripts" $sqlscripts
+    cvs tag v$tag $sqlscripts
+    echo All SQL creation scripts for release $ver generated and tagged
+  fi
+  rm -rf /tmp/cvsdiffs
+  echo
 fi
-rm -rf /tmp/cvsdiffs
-echo
